moved the logger to utils
