+++ npm/runners/build-dashboard-runner.ts
@@ -3,7 +3,6 @@ import { start } from "../scripts/dashboard/build-dashboard";
 import { dirname, resolve } from 'path';
 import { fileURLToPath } from 'url';
 
-
 const currentFilePath = fileURLToPath(import.meta.url);
 const currentDirPath = dirname(currentFilePath);
