+++ npm/runners/build-newsroom-videos-runner.ts
@@ -2,8 +2,6 @@ import { buildNewsroomVideos } from "../scripts/build-newsroom-videos";
 import { dirname, resolve } from 'path';
 import { fileURLToPath } from 'url';
 
-
-
 const currentFilePath = fileURLToPath(import.meta.url);
 const currentDirPath = dirname(currentFilePath);
